Did several gui speed improvements. Using Google API java client.

+++ src/main/java/de/chaosfisch/services/ExtendedPlaceholders.java
@@ -149,7 +149,7 @@ public class ExtendedPlaceholders {
 	 *
 	 * @return the filled number string
 	 */
-	private String zeroFill(final int number, final int width) {
+	private String zeroFill(final long number, final int width) {
 		return String.format(String.format("%%0%dd", width), number);
 	}
